requirements.txt
flask==3.1.3
flask-caching==2.3.1
flask-limiter==4.1.1
flask-sqlalchemy==3.1.1
flask-marshmallow==1.2.1
marshmallow<4
marshmallow-sqlalchemy==1.1.0
mysql-connector-python==9.6.0
python-jose==3.5.0
